Move diff package to top-level go directory
Towards #609.
