+++ cmd/noms/noms_diff.go
@@ -7,10 +7,10 @@ package main
 import (
 	"fmt"
 
-	"github.com/attic-labs/noms/cmd/noms/diff"
 	"github.com/attic-labs/noms/cmd/util"
 	"github.com/attic-labs/noms/go/config"
 	"github.com/attic-labs/noms/go/d"
+	"github.com/attic-labs/noms/go/diff"
 	"github.com/attic-labs/noms/go/util/outputpager"
 	"github.com/attic-labs/noms/go/util/verbose"
 	flag "github.com/juju/gnuflag"
