+++ go/diff/diff.go
