+++ go/diff/print_diff.go
