+++ go/diff/summary.go
